Unical course project
The program is basically a calculator that allows you to input numbers
with units. It will automatically convert the units and output the
answer in either the unit you want or in the base unit. Meanwhile, you
are also allowed to define new units as along as you define them in the
right way. All the existing units data is in QuantityDB.java, expanding
the database would be as easy as putting some extra lines in that file.